wayland: Add keyboard support
This adds basic keyboard support to the wayland compositor.
Auto-repeated key events are filtered out because the Wayland clients
are expected to generate them themselves. The wayland keyboard input
focus surface is updated whenever Mutter sees a FocusIn event so that
it will stay in synch with whatever surface Mutter wants as the focus.