wayland: Add an actor for the cursor
When running Mutter under Cogl's KMS backend no cursor will be
provided so instead this makes it so the cursor will be painted as a
ClutterTexture that gets moved in respose to mouse motion events. The
painting is done in a subclass of ClutterStage so that we can
guarantee that the actor will be painted on top of all other actors.

This patch adds support for the attach method on the input device
interface so that clients can change the cursor image.

The attach method sets a buffer and a hotspot position to use for the
cursor image. Mutter now just converts the buffer to a Cogl texture
and sets it on the ClutterTexture for the cursor. The cursor reverts
back to the default image whenever to the pointer focus is moved off
of any surface.

The image for the pointer is taken from X. It gets installed into
a fixed data location for mutter.