Allow potentially supporting multiple MetaWindowActors per window
A lot of the code in mutter assumed a one to one relationship between a
MetaWindow and a MetaWindowActor, and that it owns its MetaWindowActor. For the
most part, we've gotten around this in GNOME Shell by creating a ClutterClone
of the MetaWindowActor. In some spots, we reparent the actor owned by Mutter,
and use it for our own purpose, which just plain isn't nice. Since the Clutter
mafia has plans to drive a bullet through ClutterClone, and since
MetaWindowActor is already a very thin wrapper around MetaShapedTexture, we
should be able to create multiple MetaWindowActors for a MetaShapedTexture.

This commit replaces the "compositor_private" property on MetaWindow with
a direct reference to a MetaWindowActor (which may go away at some point, see
below), and a list of window actors.

The direct reference is a replacement for the "compositor_private" property
that has been there previously. It doesn't make sense to run mutter without
compositing, so there's no real reason to keep the veil of a compositor/core
split.

The eventual plan I have in mind is to remove the direct reference to
MetaWindowActor, and have the compositor/plugin just create a new window actor
like anybody else.