wayland: Use a signal emission hook to catch grabbed events
The wayland input device needs to see all pointer events in order to
update the cursor position. Previously it would do this just by
listening for the captured event signal on the stage. Gnome Shell sets
a pointer grab on some actors while dragging and this causes the
captured event signal not to be emitted so the cursor position does
not get updated. To workaround that this patch adds a signal emission
hook on the event signal. When a pointer or keyboard event is seen and
there is a pointer or keyboard grab in place then it will forward the
event on to the captured event handler.