wayland: Rework surface management
wl_buffer already has reference counting, so we have no need for
MetaWaylandBuffer, and can use struct wl_buffer directly.

This change also ports wl_surface_interface to the new API except for
some unimplemented parts and the fact that transactions are completely
missing so far.

Most of the xserver parts were commented out, and needs to be dealt with
later.