wayland: Paint proper pointer cursor size on HiDPI monitors
When a poiner is on a HiDPI monitor, scale the pointer sprite according
to the monitor scale. This currently assumes that the system is
configured to specify the theme size in standard DPI (meaning the theme
should assume the output scale is 1).

This works by making MetaCursorSprite more state aware during painting so
that it, given what monitor it is on and the output scale of that
monitor, can scale surface based surfaces accordingly, or recreate
server side cursor images by loading a larger version.

The native cursor renderer is changed to store up to two gbm_bo's per
MetaCursorSprite in order to draw a larger version while showing the
previous one.

https://bugzilla.gnome.org/show_bug.cgi?id=744932