Put clutter in tree and don't link with external clutter library
Having mutter and clutter separated with a stable API designed for
client applications drawing scenes on the screen while abstracting the
backend and window system is inconvenient, because mutter's needs are
very different from regular clients, and refactorizations needed for
mutter style usage is completely irrelevant for all other backends.

In order to be able to do more significant refactorizations in clutter,
fork clutter and put it inside mutter and build it into a
libmutter-clutter.so.

This is the first step. After this, also cogl will be forked and built
into a libmutter-cogl.so, and how mutter, clutter and cogl interact
will change in order to meet necessary needs regarding KMS interaction
and other display server kind of things.