wayland: Add get_parent() vfunc to the role class
What the parent of a surface is depends on the role. Here "parent" is
strictly a persistant parent child relationship, such as the one
between the popup and the popup parent, and the subsurface and the
subsurface parent.

Stacking order parent child relationship is not meant to have an effect
on the return value of this. The meta_wayland_surface_get_toplevel()
function is changed to use this new API instead of peeking at the
role dependent fields.