wayland/xdg-shell: Handle xdg_shell being destroyed too early
When clients miss behave or when a client is destroyed, xdg_shell may
be destructed before its xdg_surface's and xdg surface based roles.

Handle this by destroying child resources prior to the xdg_shell being
fully destroyed, and send errors to clients. If the reason for
unordered destruction is miss behaving clients, then the error will be
sent, if the client is already destroyed, the errors will just be
dropped on the floor.

This fixes a crash when the xdg_shell is destroyed before a popup tree,
and when the parent of a popup is destroyed before the popup. When the
popup parent is destroyed before the popup, an error would be raised,
which would cause a segmentation fault if the xdg_shell had already
been destroyed.