ClutterText: generate resource-scaled text and paint it at proper scaling
When resource scale is set we need to generate a scaled PangoLayout (by adding
a new attribute), then it has to be painted with proper scaling matrix.

So everything that has to do with PangoLayout has to be in real coordinates,
so clutter logical coords multiplied by resource scaling.
While the actual size of the layout is the one of the PangoLayout divided by
resource scale.